refactored code to maintain consistancy
