+++ conf/remote_url_conf.py
@@ -5,3 +5,4 @@ browserstack_app_upload_url = "https://api-cloud.browserstack.com/app-automate/u
 saucelabs_url = "https://ondemand.eu-central-1.saucelabs.com:443/wd/hub"
 saucelabs_app_upload_url = 'https://api.eu-central-1.saucelabs.com/v1/storage/upload'
 lambdatest_url = "http://{}:{}@hub.lambdatest.com/wd/hub"
+lambdatest_api_server_url = "https://api.lambdatest.com/automation/api/v1"
