+++ env_remote
@@ -1,5 +1,6 @@
 #Set REMOTE_BROWSER_PLATFROM TO BS TO RUN ON BROWSERSTACK else
-#SET REMOTE_BROWSER_PLATFORM TO SL TO RUN ON SAUCE LABS
+#SET REMOTE_BROWSER_PLATFORM TO SL TO RUN ON SAUCELABS
+#SET REMOTE_BROWSER_PLATFORM TO LT TO RUN ON LAMBDATEST
 REMOTE_BROWSER_PLATFORM = "BS"
 
 REMOTE_USERNAME = "Enter your username"
